+++ examples/presentation/worldLines.go
@@ -28,8 +28,8 @@ func (w *WorldLines) Create() {
 	rect := resolv.NewRectangle(160, 16, 32, 32)
 	w.Space.Add(rect)
 
-	var lx, ly int32 = 160, 160
-	var ls int32 = 16
+	var lx, ly float32 = 160, 160
+	var ls float32 = 16
 
 	line := resolv.NewLine(lx, ly, lx+ls, ly)
 	w.Space.Add(line)
@@ -56,8 +56,8 @@ func (w *WorldLines) Update() {
 	ratioX := float32(screenWidth) / float32(winW)
 	ratioY := float32(screenHeight) / float32(winH)
 
-	mx := int32(float32(x) * ratioX)
-	my := int32(float32(y) * ratioY)
+	mx := float32(float32(x) * ratioX)
+	my := float32(float32(y) * ratioY)
 
 	if rl.IsMouseButtonPressed(rl.MouseLeftButton) {
 		w.TargetLine.X = mx
@@ -82,8 +82,8 @@ func (w *WorldLines) Draw() {
 			if line == w.TargetLine {
 				if w.Space.IsColliding(line) {
 					for i, point := range line.GetIntersectionPoints(w.Space) {
-						rl.DrawLine(point.X-5, point.Y-5, point.X+5, point.Y+5, rl.Yellow)
-						rl.DrawLine(point.X+5, point.Y-5, point.X-5, point.Y+5, rl.Yellow)
+						rl.DrawLine(int32(point.X-5), int32(point.Y-5), int32(point.X+5), int32(point.Y+5), rl.Yellow)
+						rl.DrawLine(int32(point.X+5), int32(point.Y-5), int32(point.X-5), int32(point.Y+5), rl.Yellow)
 						DrawText(point.X+5, point.Y, "Intersection #"+strconv.Itoa(i+1))
 					}
 					drawColor = rl.Red
@@ -91,7 +91,7 @@ func (w *WorldLines) Draw() {
 					drawColor = rl.Green
 				}
 			}
-			rl.DrawLine(line.X, line.Y, line.X2, line.Y2, drawColor)
+			rl.DrawLine(int32(line.X), int32(line.Y), int32(line.X2), int32(line.Y2), drawColor)
 
 		}
 
@@ -99,7 +99,7 @@ func (w *WorldLines) Draw() {
 
 		if ok {
 
-			rl.DrawRectangleLines(rect.X, rect.Y, rect.W, rect.H, rl.LightGray)
+			rl.DrawRectangleLines(int32(rect.X), int32(rect.Y), int32(rect.W), int32(rect.H), rl.LightGray)
 
 		}
